refactored authorization and authentication packages
